+++ app/include/tab/setting_tab.hpp
@@ -24,11 +24,14 @@ class SettingTab : public AttachedView {
 public:
     SettingTab();
 
+    void hideStatus();
+
     void onCreate() override;
 
     static brls::View* create();
 
 private:
+    BRLS_BIND(brls::Box, boxStatus, "setting/status");
     BRLS_BIND(brls::RadioCell, btnTutorialOpenApp, "tools/tutorial_open");
     BRLS_BIND(brls::RadioCell, btnTutorialError, "tools/tutorial_error");
     BRLS_BIND(brls::RadioCell, btnTutorialFont, "tools/tutorial_font");