+++ app/src/tab/server_add.cpp
@@ -4,6 +4,7 @@
 
 #include "tab/server_add.hpp"
 #include "tab/server_login.hpp"
+#include "tab/setting_tab.hpp"
 #include "utils/config.hpp"
 #include "utils/dialog.hpp"
 #include "api/jellyfin.hpp"
@@ -14,6 +15,15 @@ ServerAdd::ServerAdd(std::function<void(void)> cb) : cbConnected(cb) {
     // Inflate the tab from the XML file
     this->inflateFromXMLRes("xml/tabs/server_add.xml");
     brls::Logger::debug("ServerAdd: create");
+
+    this->registerAction(
+        "main/tabs/setting"_i18n, brls::BUTTON_Y, [this](...) {
+            SettingTab *view = new SettingTab();
+            view->hideStatus();
+            view->onCreate();
+            this->present(view);
+            return true;
+        });
 
     inputUrl->init("URL", "https://", [](std::string) {}, "", "", 255);