+++ app/src/tab/setting_tab.cpp
@@ -76,32 +76,36 @@ SettingTab::SettingTab() {
     GA("open_setting");
 }
 
+void SettingTab::hideStatus() { boxStatus->setVisibility(brls::Visibility::GONE); }
+
 void SettingTab::onCreate() {
     auto& conf = AppConfig::instance();
 
-    btnServer->setDetailText(conf.getUrl());
-    btnServer->registerClickAction([](...) {
-        brls::Application::pushActivity(new ServerList(), brls::TransitionAnimation::NONE);
-        return true;
-    });
-
-    btnUser->setDetailText(conf.getUser().name);
-    btnUser->registerClickAction([](...) {
-        Dialog::cancelable("main/setting/others/logout"_i18n, []() {
-            brls::async([]() {
-                auto& c = AppConfig::instance();
-                HTTP::Header header = {"X-Emby-Token: " + c.getUser().access_token};
-                try {
-                    HTTP::post(c.getUrl() + jellyfin::apiLogout, "", header, HTTP::Timeout{});
-                    c.removeUser(c.getUser().id);
-                } catch (const std::exception& ex) {
-                    brls::Logger::warning("Logout failed: {}", ex.what());
-                }
-                brls::sync([]() { brls::Application::quit(); });
-            });
+    if (boxStatus->getVisibility() == brls::Visibility::VISIBLE) {
+        btnServer->setDetailText(conf.getUrl());
+        btnServer->registerClickAction([](...) {
+            brls::Application::pushActivity(new ServerList(), brls::TransitionAnimation::NONE);
+            return true;
         });
-        return true;
-    });
+
+        btnUser->setDetailText(conf.getUser().name);
+        btnUser->registerClickAction([](...) {
+            Dialog::cancelable("main/setting/others/logout"_i18n, []() {
+                brls::async([]() {
+                    auto& c = AppConfig::instance();
+                    HTTP::Header header = {"X-Emby-Token: " + c.getUser().access_token};
+                    try {
+                        HTTP::post(c.getUrl() + jellyfin::apiLogout, "", header, HTTP::Timeout{});
+                        c.removeUser(c.getUser().id);
+                    } catch (const std::exception& ex) {
+                        brls::Logger::warning("Logout failed: {}", ex.what());
+                    }
+                    brls::sync([]() { brls::Application::quit(); });
+                });
+            });
+            return true;
+        });
+    }
 
 /// Hardware decode
 #ifdef __SWITCH__